Add script to get list of used source files
This script uses the given .config file and make arguments to build
the kernel in a temporary directory.  It sets the AS, CC, and HOSTCC
make variables to interpose a script that records all the used C and
assembly sources.

It then adds all other files in the repository that might
(conservatively) be needed to build the kernel.  It excludes
configuration files, C and assembly sources, documentation, and
user-space tools other than build tools.

My initial approach was to use the checker hook (C=2 CHECK=...), but
it is not invoked for assembly sources or the C sources for build
tools (kconfig, modpost, objtool, etc.).  This makes it difficult to
both include all used sources and avoid including large numbers of
unused sources.