get_used_sources.py: Remove references to missing files from kernel config
Various kconfig symbols can refer to extra files that are not part of
the upstream source, such as firmware blobs and certificates.  Since
we don't have those files in the kernel-config repository, such
references will cause build failures.

Remove these references when copying the .config file, and warn when
we do so as this may cause the sources of some build tools to be
omitted.