Add script to check and show warnings for configuration files
For the 4.4 branches I already identified some unsupportable kernel
features and configurations that I recommended disabling.  I will
shortly do the same for the 4.19 branches.

Rather than writing this simply as free text, I think it's preferable
to at least summarise the problem features in a machine-readable
format that can be checked automatically.  Define that format and add
a script to do the checking.

Signed-off-by: Ben Hutchings <ben.hutchings@codethink.co.uk>